Img_workflow/requirements.txt
fastapi
uvicorn
pandas
openpyxl
requests
python-dotenv
gspread
google-genai
Pillow
supabase
